Switch to Github Packages

+++ src/test/java/org/ligoj/app/plugin/prov/google/ProvGoogleResourceTest.java
@@ -86,7 +86,7 @@ public class ProvGoogleResourceTest extends AbstractAppTest {
 		Assertions.assertNotNull(instance.getId().intValue());
 		Assertions.assertEquals("f1-micro", instance.getName());
 		Assertions.assertEquals(0.2, instance.getCpu(), 0.01);
-		Assertions.assertEquals(614, instance.getRam().intValue());
+		Assertions.assertEquals(614, instance.getRam());
 		Assertions.assertFalse(instance.getConstant());
 
 		Assertions.assertEquals("SQL Server Enterprise", instances.get(2).getPrice().getLicense());
